Adds a few more utilities to reactive.dom
Warning: this is an API breaking change!  Renames bindToInput to bindToChange
and adds a new bindToInput.  Each of these bind to the respective jQuery event
of their namesake ('input' and 'change' respectively).  Also consistently forces
an initial evaluation of the bound function so it begins in an initialized state
(rather than only being called when the event triggers the first time).

Also adds two new utilities:
-bindPropertyToEach for binding to many a function's output to many
elements in a single call.
-bindToggleClassTo for binding a functions output to the toggleClass jQuery
method so a class may be toggled (not just added).